mini_memo: fix CMake build by adding ai_agent include paths
mini_memo_core.c includes <velaclaw/client.h>, but only the Makefile
carried the ai_agent include paths -- the CMake target had no matching
INCLUDE_DIRECTORIES, so building with CMake failed on "velaclaw/client.h:
No such file or directory". Pass the same two paths to
nuttx_add_application.

Also fix the version log format: cJSON_int is long long in this
configuration, so "%d" tripped -Werror=format= and broke the build even
after the include paths were resolved. Use %lld with an explicit cast.

+++ mini_memo/mini_memo_core.c
@@ -328,8 +328,8 @@ int memo_store_load(void)
     }
 
     if (version_obj->valueint != MEMO_JSON_VERSION) {
-        syslog(LOG_WARNING, "%s: unsupported version %d\n",
-            MEMO_TAG, version_obj->valueint);
+        syslog(LOG_WARNING, "%s: unsupported version %lld\n",
+            MEMO_TAG, (long long)version_obj->valueint);
         cJSON_Delete(root);
         root = NULL;
         return 0;
